adds refactoring for multiple reducers in redux
