+++ src/redux/reducers/index.js
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux';
+
+import transport from './transport';
+import mixer from './mixer';
+
+export default combineReducers({
+  transport,
+  mixer,
+});
