+++ src/redux/store.js
@@ -1,5 +1,5 @@
 import { createStore } from 'redux';
-import reducer from './reducer';
+import reducer from './reducers';
 
 const store = createStore(reducer);
